+++ src/bmp/bmp.c
@@ -211,7 +211,7 @@ void skinny_bmp_daemon()
 
   if (config.nfacctd_bmp_pipe_size) {
     int l = sizeof(config.nfacctd_bmp_pipe_size);
-    u_int64_t saved = 0, obtained = 0;
+    int saved = 0, obtained = 0;
 
     getsockopt(config.bmp_sock, SOL_SOCKET, SO_RCVBUF, &saved, &l);
     Setsocksize(config.bmp_sock, SOL_SOCKET, SO_RCVBUF, &config.nfacctd_bmp_pipe_size, sizeof(config.nfacctd_bmp_pipe_size));
@@ -219,7 +219,7 @@ void skinny_bmp_daemon()
 
     Setsocksize(config.bmp_sock, SOL_SOCKET, SO_RCVBUF, &saved, l);
     getsockopt(config.bmp_sock, SOL_SOCKET, SO_RCVBUF, &obtained, &l);
-    Log(LOG_INFO, "INFO ( %s/core/BMP ): bmp_daemon_pipe_size: obtained=%u target=%u.\n", config.name, obtained, config.nfacctd_bmp_pipe_size);
+    Log(LOG_INFO, "INFO ( %s/core/BMP ): bmp_daemon_pipe_size: obtained=%d target=%d.\n", config.name, obtained, config.nfacctd_bmp_pipe_size);
   }
 
   rc = bind(config.bmp_sock, (struct sockaddr *) &server, slen);
